requirements.txt
python-telegram-bot==20.7
ccxt==4.2.78
pandas==2.1.4